Struct crate_metadata_serde::SerializedCrate

pub struct SerializedCrate {
    pub crate_name: String,
    pub sections: HashMap<Shndx, SerializedSection>,
    pub global_sections: BTreeSet<Shndx>,
    pub tls_sections: BTreeSet<Shndx>,
    pub data_sections: BTreeSet<Shndx>,
    pub init_symbols: BTreeMap<String, usize>,
}

A (de)serializable representation of a loaded crate that is serde-compatible.

See LoadedCrate for more detail on the fields of this struct.

Fields
crate_name: String

The name of the crate.

sections: HashMap<Shndx, SerializedSection>

A map containing all the sections of the crate.

global_sections: BTreeSet<Shndx>

A set containing the global sectinos of the crate.

tls_sections: BTreeSet<Shndx>

A set containing the thread-local storage (TLS) sections of the crate.

data_sections: BTreeSet<Shndx>

A set containing the .data and .bss sections of the crate.

init_symbols: BTreeMap<String, usize>

A map of symbol names to their constant values, which contain assembler and linker constants.